# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise Platform REST API",
    "version": "0.1.0",
    "description": "REST supplementary routes for the enterprise React platform BFF. GraphQL is the primary client boundary per ADR-0013; these routes cover health probes, auth, session management, FQDN-based tenant routing, per-tenant branding, auth settings self-service, and tenant provisioning. ADRs 0028\u20130031 govern the multi-tenant architecture."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev (platform-api)"
    },
    {
      "url": "https://aldous.info",
      "description": "Production (Cloudflare)"
    }
  ],
  "paths": {
...
    "/api/session": {
      "get": {
        "summary": "Get current session actor",
        "operationId": "getSession",
        "tags": ["Auth"],
        "responses": {
          "200": {
            "description": "Authenticated session actor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionActor"
                }
              }
            }
          },
          "401": {
            "description": "No active session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/graphql": {
      "post": {
        "summary": "GraphQL application boundary",
        "operationId": "graphql",
        "tags": ["GraphQL"],
        "responses": {
          "200": {
            "description": "GraphQL response"
          },
          "400": {
            "description": "Invalid GraphQL query or operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/logs/search": {
      "get": {
        "summary": "Search operator logs",
        "operationId": "adminLogsSearch",
        "tags": ["Admin Logs"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Free-text search query"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            },
            "description": "Maximum number of entries to return"
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["forward", "backward"],
              "default": "backward"
            },
            "description": "Pagination direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching log entries"
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/settings/idps": {
      "get": {
        "summary": "List tenant identity providers",
        "operationId": "listIdps",
        "tags": ["Auth Settings"],
        "description": "Returns configured IdPs for the tenant realm derived from the Host header. Requires admin.access permission. ADR-0030 \u00a71b.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "IdP list"
          },
          "400": {
            "description": "No tenant context"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "summary": "Create or update identity provider",
        "operationId": "upsertIdp",
        "tags": ["Auth Settings"],
        "description": "Creates or updates an IdP in the tenant Keycloak realm. Realm derived from Host header \u2014 never from body. ADR-0030 \u00a71b.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["alias", "displayName", "providerId"],
                "properties": {
                  "alias": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "providerId": {
                    "type": "string",
                    "enum": ["oidc", "saml", "keycloak-oidc"]
                  },
                  "config": {
                    "type": "object"
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/auth/settings/mfa": {
      "get": {
        "summary": "Get MFA policy",
        "operationId": "getMfaPolicy",
        "tags": ["Auth Settings"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "MFA policy"
          }
        }
      },
      "patch": {
        "summary": "Set MFA policy",
        "operationId": "setMfaPolicy",
        "tags": ["Auth Settings"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["required"],
                "properties": {
                  "required": {
                    "type": "string",
                    "enum": ["none", "optional", "required"]
                  },
                  "type": {
                    "type": "string",
                    "enum": ["totp", "webauthn"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/auth/settings/session": {
      "get": {
        "summary": "Get session policy",
        "operationId": "getSessionPolicy",
        "tags": ["Auth Settings"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Session policy"
          }
        }
      },
      "patch": {
        "summary": "Set session policy",
        "operationId": "setSessionPolicy",
        "tags": ["Auth Settings"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated"
          }
        }
      }
    },
...
    "/api/org/members": {
      "get": {
        "summary": "List tenant members and pending invitations",
        "operationId": "org.members.list",
        "tags": ["Members"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Member list"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/org/members/invite": {
      "post": {
        "summary": "Invite a member",
        "operationId": "org.members.invite",
        "tags": ["Members"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Invitation sent or member added"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/org/members/:userId": {
      "patch": {
        "summary": "Update member role",
        "operationId": "org.members.update_role",
        "tags": ["Members"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Role updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          },
          "422": {
            "description": "Last admin cannot be demoted"
          }
        }
      },
      "delete": {
        "summary": "Remove member",
        "operationId": "org.members.remove",
        "tags": ["Members"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          },
          "422": {
            "description": "Last admin cannot be removed"
          }
        }
      }
    },
...
    "/api/org/features": {
      "get": {
        "summary": "List feature toggle states",
        "operationId": "org.features.list",
        "tags": ["Features"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Feature states"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/org/features/:featureKey": {
      "patch": {
        "summary": "Toggle a feature",
        "operationId": "org.features.toggle",
        "tags": ["Features"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "featureKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Toggle applied"
          },
          "400": {
            "description": "Invalid body"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Unknown feature key"
          }
        }
      }
    },
...
    "/api/auth/settings/providers": {
      "get": {
        "summary": "Get tenant authentication provider config",
        "operationId": "getAuthProviders",
        "tags": ["Auth Settings"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant authentication provider config"
          }
        }
      },
      "patch": {
        "summary": "Update tenant authentication provider config",
        "operationId": "setAuthProviders",
        "tags": ["Auth Settings"],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": ["mock", "real", "disabled", "default"]
                  },
                  "enabledProviders": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": ["google", "azure", "apple", "platform"]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated config"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/org/webhooks/readiness": {
      "get": {
        "summary": "Get tenant webhooks readiness",
        "operationId": "getOrgWebhooksReadiness",
        "description": "Returns the classified webhooks readiness (no_subscriptions / configured / degraded). Never faked. ADR-0051.",
        "tags": ["Organisation"],
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Readiness status" },
          "400": { "description": "No tenant context" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "platform_session"
      }
    },
    "schemas": {
...
      "SessionActor": {
        "type": "object",
        "required": ["userId", "tenantId", "organisationId", "roles", "permissions", "displayName"],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string"
          },
          "organisationId": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
    },
...
  }
}